requirements.txt
streamlit==1.25.0
streamlit-chat==0.0.2
langchain==0.0.235
python-dotenv==1.0.0
openai==0.27.10
